OLMIS-663 Integration testing of Service & DB

+++ src/main/java/org/openlmis/example/repository/NotificationRepository.java
@@ -3,5 +3,5 @@ package org.openlmis.example.repository;
 import org.openlmis.example.domain.Notification;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NotificationRepository extends CrudRepository<Notification, Long> {
+public interface NotificationRepository extends CrudRepository<Notification, Integer> {
 }
